requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-dotenv
APScheduler
httpx
tenacity
pydantic
google-generativeai
